+++ src/utils/utils2.c
@@ -3,7 +3,7 @@
 
 int has_entity(char **map, int entity)
 {
-	int	location[2];
+	double	location[2];
 
 	find_item_pos(map, entity, &location[0], &location[1]);
 	if (location[0] == -1)
